Fixing DVI detection code crash
The portion of the code responsible for detecting VT1632A external TMDS
transmitter for DVI was causing a crash if VT1632A was not detected on
I2C buses. This was happening due to xf86DestroyI2CDevRec being called
prior to the call to xf86I2CDevInit if VT1632A was not detected. Now the
code will probe I2C buses using xf86I2CProbeAddress to look for VT1632A,
and if VT1632A is not detected, it will stop trying to detect VT1632A
via I2C buses.

Signed-off-by: Kyle Guinn <elyk03@gmail.com>
Signed-off-by: Kevin Brace <kevinbrace@gmx.com>