Compilation hooks to support ClockRange and ClockRanges
When compiling OpenChrome for Ubuntu 10.04 LTS, it was discovered that
x.org xserver discontinued the use of a structure with ClockRanges type
within ScrnInfoRec. XORG_VERSION_NUMERIC will now be used to figure out
which version of xserver the compiler is compiling against. Now
OpenChrome can be compiled against Ubuntu 10.04 LTS without warnings or
errors.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>