Using I2C bus 2 to detect a VGA monitor
Previously, it was assumed that I2C bus 1 is used to detect only a VGA
monitor. It was also assumed that I2C bus 2 is used to detect a DVI
monitor or TV encoder. However, for devices with a DVI-I connector,
VGA signals come out of the connector as well. What this means is that
I2C bus 2 also has to be used to detect a VGA monitor, in addition to
I2C bus 1. Furthermore, EDID obtained from the monitor via I2C bus
will be used to determine whether or not it is an analog type
(i.e., VGA monitor).

Reported-by: Christopher Havel <laserhawk64@gmail.com>
Tested-by: Justin Chevrier <jchevrier@gmail.com>
Signed-off-by: Kevin Brace <kevinbrace@gmx.com>