Fixing the misdetection of NanoBook flat panel
It was observed in a VIA Technologies NanoBook based laptop computer
that the flat panel it contains is not connected to an I2C bus, hence
the supported screen resolution cannot be determined via an I2C bus.
Furthermore, the code was obtaining EDID information via I2C bus 2, but
it was not scrutinizing the interface type. Hence, if a VGA monitor
is connected to the computer, the code was determining the flat panel
screen resolution based on the EDID information obtained from the VGA
monitor via I2C bus 2. Due to all of these issues, the flat panel
screen color was going completely white, making the flat panel
completely useless. Thanks to this fix, the flat panel and a VGA
monitor can now be detected correctly.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>