Removing the known device table
The code has had a large known device table that was used to aid the
display device detection for a long time. This is really an awkward way
to decide which display device gets activated. The general direction of
the computing industry is towards automatic detection of devices, hence,
the use of the known device table will be discontinued for good.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>