Fix for screen resolution change crash and extended screen support
Not checking the validity of an old screen pointer before freeing the
allocated memory was leading to a crash of X.org X Server in most
cases when the screen resolution was changed. With this fix, extended
screen using 2 monitors appears to be working in some cases. This is
a major bug fix.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>