Version reverted back to 0.4.102
Version 0.5.0 RC1 (Release Candidate 1) is canceled due to a bug where
the IGA1 screen freezes when IGA2 is turned off and screen resolution
of IGA1 is changed from the OS. This bug was discovered during testing
after the release of Version 0.5.0 RC1. There needs to be a bug fix,
hence, this version is no longer appropriate to be an RC (i.e., Patch
Level >= 900). The last Patch Level prior to 900 was 101, so 102 is now
the new Patch Level.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>