Fix a fatal regression impacting runtime screen resize
It was discovered that a commit done between Version 0.3.2
and Version 0.3.3 broke the screen resize functionality
(Commit cee0a1fab9cade87e6de16c67cd34c84cf697531). This led to
an X Server crash if the screen was resized from a screen resize
utility. With this fix, the screen resolution can now be changed
from the screen resize utility safely.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>