VT1632A is finally working with VX855 chipset!
The VT1632A register settings might be too specific to Wyse C00X
thin client, but with this commit, DVI is now working with that unit.
VT1632A is connected to DVP1 (Digital Video Port 1), and the device
driver is (somehow) using IGA2 for DVI. It is possible that one might
be able to do use 2 monitors (VGA and DVI) simultaneously in a
non-clone mode. This also means that the current code might also work
with CX700, VX700, VX800, and VX900 chipsets.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>