Fix for xf86InitFBManager implicit declaration warning
Commit e73fa19 removed including xf86fbman.h when compiling
via_ums.c, and as a result, if the code was compiled with a
stricter warning compilation option, it will not be able to
successfully compile the code due to it not being able to
resolve the lack of xf86InitFBManager declaration. Including
xf86fbman.h with via_driver.h resolves this problem.

Reported-by: Benno Schulenberg <bensberg@justemail.net>
Tested-by: Xavier Bachelot <xavier@bachelot.org>
Signed-off-by: Kevin Brace <kevinbrace@gmx.com>