VIAFreeRec will be called from VIAFreeScreen
Prior to commit 870d3c7, VIAFreeRec function was being called prior
to the check to see if calling vgaHWFreeHWRec is necessary.
Unfortunately, this check was referencing DDX's private storage area
after it was freed by the call to VIAFreeRec function. This was
leading to a segmentation fault during the X Server shutdown. In order
to remedy this, the call to VIAFreeRec function was removed by commit
870d3c7. However, according to XFree86 DDX Design document
(ddxDesign.pdf), VIAFreeRec function is expected to be called from
VIAFreeScreen callback function. Not doing so very likely will lead to
memory leak, hence, VIAFreeRec function will now be called from
VIAFreeScreen callback function, but after vgaHWFreeHWRec is called.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>