Removal of questionable lines of code from viaIGA1SetDisplayRegister
It was access to SR1A (3C5.1A) that was key to this portion of
the code. Removing this line led CLE266 and KM400 to freeze right
after X.Org Server booted. It appears that setting SR1A bit 6 to 0
resolves the bug. This bit performs reset of VIA IGP from software.
It is likely that this bit was set to 1 in CLE266 and KM400 chipsets
for some reason. Now viaIGAInitCommon function handles the zero
clearing of this bit.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>