Initial support for integrated TMDS transmitter (DVI)
This commit incorporates the initial support for VIA Technologies
IGP integrated TMDS transmitter. Previously, the code to support
the integrated TMDS transmitter was reusing the code written for
handling integrated LVDS transmitters. This was causing number of
problems, so totally new code was written to support the integrated
TMDS transmitter only. The integrated TMDS transmitter identifies
itself as DVI (Digital Visual Interface). The code was written in a
way to coexist with the existing code that initializes VIA
Technologies VT1632(A) external TMDS transmitter.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>