Designating DVI dynamically for xrandr
The code was updated so that DVI designation for xrandr will have
a numbering convention that will look like, DVI-1, DVI-2, etc. If
both integrated and external TMDS transmitters are detected in the
same system, the integrated one will be assigned DVI-1 first. The
code to facilitate this was borrowed from xf86-video-r128 DDX.

Suggested-by: Eric Kudzin <knoppix1337@yahoo.com>
Signed-off-by: Kevin Brace <kevinbrace@gmx.com>