Silicon Image SiI 164 initial code inclusion
This is the initial code inclusion to support Silicon Image SiI 164
external TMDS transmitter. The code to support SiI 164 was borrowed
from the proven VIA Technologies VT1632(A) external TMDS transmitter
code. The code is not likely to work at this point since it has not
been tested due to lack of equipment availability.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>