Giving the original author of external TMDS transmitter code credit
Luc Verhaegen notified that the code to support Silicon Image SiI 164
and VIA Technologies VT1632(A) closely resembles the code he wrote
for xf86-video-unichrome. After examining the code, portions of it
appears to resemble Luc's original code. Christian Jung who was the
author of the VIA Technologies VT1632(A) code acknowledged the lack
of Luc's copyright in the code he submitted.

Suggested-by: Luc Verhaegen <libv@skynet.be>
Acked-by: JUNG, Christian <christian.jung@shsservices.org>
Signed-off-by: Kevin Brace <kevinbrace@gmx.com>