Support for automatic TV encoder display source assignment
Code was added to perform automatic display source assignment
(i.e., IGA1 or IGA2) when an external TV encoder is attached. The
code was written in a way it can detect which external interface
has the TV encoder attached to it, and select the correct external
interface automatically. The code supports all Chrome IGPs. (from
CLE266 chipset all the way to VX900 chipset)

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>