Support for automatic TV encoder I/O pad turn on
Code was added to perform automatic TV encoder I/O pad turn on.
The code was written in a way it can detect which external interface
has the TV encoder attached to it, and select the correct external
interface automatically. The code supports all Chrome IGPs. (from
CLE266 chipset all the way to VX900 chipset)

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>