Permanent fix for HP 2133 PCIe WLAN not working
Using strapping pins for Chrome IGP help figure out whether or not to
activate DVP0. Tested the fix with 3 different VN896 chipset based
laptops, but all of these have a 12-bit FP interface. The fix is untested
with a 24-bit FP interface. This fix also helps to restore the FP
correctly when coming out of standby in HP 2133 Mini-Note.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>