Faking integrated TMDS (DVI) transmitter DVI detection
Probably not a good idea, but do not have a choice for now.
Integrated TMDS transmitter based DVI's detection pin does not really
work, so this is how detection needs to be faked. This issue does not
affect those where the DVI is coming from an external TMDS transmitter
like SiI 164 or VT1632(A).

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>