Version bumped to 0.6.121
This version fixes a bug exposed by the updated analog (VGA) and
FP initialization code. Thanks to the updated code, analog and FP
can be assigned to IGA1 or IGA2. However, IGA1 DPMS control code
was controlling analog DPMS, and this led to a buggy behavior
of analog display when turning off analog display from the OS in
some situations. The analog DPMS now correctly controls the DPMS
state, and IGA1 output control is done differently.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>