Fix for DVI (integrated TMDS transmitter) power management
Due to the recent major updates to the FP power management code, the
existing DVI power management code was no longer able to manage the
power state of DVI correctly. This commit fixes this regression. This
commit affects CX700 / VX700 and VX800 chipsets that utilize the
integrated TMDS transmitter for DVI use.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>