VT switch fix for drm-openchrome
The changes made to the code with commit 7095fd0 was leading to
a null pointer crash. This crash will take down the X Server all
together. This commit was made in order to accommodate Linux 4.5
or later’s change in memory mapping related to vesafb. In reality,
the code itself was not separating legacy mode setting (UMS) and
DRM based mode setting (KMS) very well prior to commit 7095fd0 to
begin with, so this additional commit is necessary. It will not be
surprising if more commits like this one will be necessary in order
to stabilize the DDX code.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>