Activate hardware cursor for CLE266 and KM400 chipsets
Unfortunately, the current HI (Hardware Icon) based overlay cursor
code does not work properly on CLE266 and KM400 chipsets.  Instead,
use bi-color hardware cursor on these platforms.  The limitation
of the code is that it can only display the hardware cursor on one
CRTC (IGA1 or IGA2).  This is due to hardware limitation.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>