Version bumped to 0.6.409
Added the necessary code for CLE266 chipset to support an external
DVI transmitter.  Both Silicon Image SiI 164 and VIA Technologies
VT1632(A) have the relevant code.  The code was tested on NeoWare
CA10 model with DVI.  It should work on VIA Embedded EPIA-M with
DVI-01 module, but it was not confirmed (do not have the DVI-01
module).  While the code was not tested with an LCD panel, some of
the added new code improves the CLE266 chipset's LCD panel support.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>